requirements.txt
Django==4.2.7
psycopg2-binary==2.9.7
python-decouple==3.8
requests==2.31.0
pytz==2023.3
django-cors-headers==4.3.1
Pillow==10.0.1
django-crispy-forms==2.1
crispy-bootstrap5==0.7
celery==5.3.4
redis==5.0.1
gunicorn==21.2.0
whitenoise==6.6.0

